rename module to `scott`
