--- curvature_filtrations/__init__.py
@@ -1,2 +1,0 @@
-from curvature_filtrations.compare import Comparator
-from curvature_filtrations.kilt import KILT
