+++ scott/__init__.py
@@ -1,0 +1,2 @@
+from scott.compare import Comparator
+from scott.kilt import KILT
