+++ scott/topology/representations.py
